LPS-67836 Rename to Social Office Associations (because it handles the associations between users, roles and orgs) and move the listeners to a package model.listener
