+++ modules/apps/social-office-upgrade/social-office-association/bnd.bnd
@@ -1,5 +1,5 @@
-Bundle-Name: Liferay Social Office Listeners
-Bundle-SymbolicName: com.liferay.social.office.listeners
+Bundle-Name: Liferay Social Office Association
+Bundle-SymbolicName: com.liferay.social.office.association
 Bundle-Version: 1.0.0
 Liferay-Releng-Module-Group-Description:
 Liferay-Releng-Module-Group-Title: Social Office Upgrade
