+++ modules/apps/social-office-upgrade/social-office-association/build.gradle
