+++ modules/apps/social-office-upgrade/social-office-association/src/main/java/com/liferay/social/office/association/internal/constants/SocialOfficeConstants.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.social.office.internal.listeners.constants;
+package com.liferay.social.office.association.internal.constants;
 
 /**
  * @author Eudaldo Alonso
